requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.7
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
httptools==0.6.4
idna==3.10
mangum==0.19.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
